requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
python-decouple==3.3
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
